risk_model/requirements.txt
# Requirements for DPR Risk Prediction System
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
transformers>=4.18.0
nltk>=3.6.0
spacy>=3.2.0
PyPDF2>=2.0.0
python-docx>=0.8.11
openpyxl>=3.0.9
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.1.0
tqdm>=4.62.0
torch>=1.10.0
flask>=2.0.0
python-dotenv>=0.19.0
pydantic>=1.9.0